Alzheimer's disease data

Alzheimer's disease data

Source

Kuhn, M., Johnson, K. (2013) Applied Predictive Modeling, Springer.

Craig-Schapiro R, Kuhn M, Xiong C, Pickering EH, Liu J, Misko TP, et al. (2011) Multiplexed Immunoassay Panel Identifies Novel CSF Biomarkers for Alzheimer's Disease Diagnosis and Prognosis. PLoS ONE 6(4): e18850.

Value

ad_data

a tibble

Details

Craig-Schapiro et al. (2011) describe a clinical study of 333 patients, including some with mild (but well-characterized) cognitive impairment as well as healthy individuals. CSF samples were taken from all subjects. The goal of the study was to determine if subjects in the early states of impairment could be differentiated from cognitively healthy individuals. Data collected on each subject included:

Demographic characteristics such as age and gender

Apolipoprotein E genotype

Protein measurements of Abeta, Tau, and a phosphorylated version of Tau (called pTau)

Proteinmeasurements of 124 exploratory biomarkers, and

Clinical dementia scores

For these analyses, we have converted the scores to two classes: impaired and healthy. The goal of this analysis is to create classification models using the demographic and assay data to predict which patients have early stages of disease.

Examples

data(ad_data)
str(ad_data)
#> tibble [333 × 131] (S3: tbl_df/tbl/data.frame)
#>  $ ACE_CD143_Angiotensin_Converti  : num [1:333] 2 1.56 1.52 1.68 2.4 ...
#>  $ ACTH_Adrenocorticotropic_Hormon : num [1:333] -1.386 -1.386 -1.715 -1.609 -0.968 ...
#>  $ AXL                             : num [1:333] 1.098 0.683 -0.145 0.683 0.191 ...
#>  $ Adiponectin                     : num [1:333] -5.36 -5.02 -5.81 -5.12 -4.78 ...
#>  $ Alpha_1_Antichymotrypsin        : num [1:333] 1.74 1.46 1.19 1.28 2.13 ...
#>  $ Alpha_1_Antitrypsin             : num [1:333] -12.6 -11.9 -13.6 -15.5 -11.1 ...
#>  $ Alpha_1_Microglobulin           : num [1:333] -2.58 -3.24 -2.88 -3.17 -2.34 ...
#>  $ Alpha_2_Macroglobulin           : num [1:333] -72.7 -154.6 -136.5 -98.4 -144.9 ...
#>  $ Angiopoietin_2_ANG_2            : num [1:333] 1.065 0.742 0.833 0.916 0.956 ...
#>  $ Angiotensinogen                 : num [1:333] 2.51 2.46 1.98 2.38 2.86 ...
#>  $ Apolipoprotein_A_IV             : num [1:333] -1.43 -1.66 -1.66 -2.12 -1.17 ...
#>  $ Apolipoprotein_A1               : num [1:333] -7.4 -7.05 -7.68 -8.05 -6.73 ...
#>  $ Apolipoprotein_A2               : num [1:333] -0.2614 -0.8675 -0.6539 -1.2379 0.0953 ...
#>  $ Apolipoprotein_B                : num [1:333] -4.62 -6.75 -3.98 -6.52 -3.38 ...
#>  $ Apolipoprotein_CI               : num [1:333] -1.273 -1.273 -1.715 -1.966 -0.755 ...
#>  $ Apolipoprotein_CIII             : num [1:333] -2.31 -2.34 -2.75 -3 -1.51 ...
#>  $ Apolipoprotein_D                : num [1:333] 2.08 1.34 1.34 1.44 1.63 ...
#>  $ Apolipoprotein_E                : num [1:333] 3.75 3.1 2.75 2.37 3.07 ...
#>  $ Apolipoprotein_H                : num [1:333] -0.157 -0.575 -0.345 -0.532 0.663 ...
#>  $ B_Lymphocyte_Chemoattractant_BL : num [1:333] 2.3 1.67 1.67 1.98 2.3 ...
#>  $ BMP_6                           : num [1:333] -2.2 -1.73 -2.06 -1.98 -1.24 ...
#>  $ Beta_2_Microglobulin            : num [1:333] 0.693 0.47 0.336 0.642 0.336 ...
#>  $ Betacellulin                    : int [1:333] 34 53 49 52 67 51 41 42 58 59 ...
#>  $ C_Reactive_Protein              : num [1:333] -4.07 -6.65 -8.05 -6.21 -4.34 ...
#>  $ CD40                            : num [1:333] -0.796 -1.273 -1.242 -1.124 -0.924 ...
#>  $ CD5L                            : num [1:333] 0.0953 -0.6733 0.0953 -0.3285 0.3633 ...
#>  $ Calbindin                       : num [1:333] 33.2 25.3 22.2 23.5 21.8 ...
#>  $ Calcitonin                      : num [1:333] 1.386 3.611 2.116 -0.151 1.308 ...
#>  $ CgA                             : num [1:333] 398 466 348 334 443 ...
#>  $ Clusterin_Apo_J                 : num [1:333] 3.56 3.04 2.77 2.83 3.04 ...
#>  $ Complement_3                    : num [1:333] -10.4 -16.1 -16.1 -13.2 -12.8 ...
#>  $ Complement_Factor_H             : num [1:333] 3.57 3.6 4.47 3.1 7.25 ...
#>  $ Connective_Tissue_Growth_Factor : num [1:333] 0.531 0.588 0.642 0.531 0.916 ...
#>  $ Cortisol                        : num [1:333] 10 12 10 14 11 13 4.9 13 12 17 ...
#>  $ Creatine_Kinase_MB              : num [1:333] -1.71 -1.75 -1.38 -1.65 -1.63 ...
#>  $ Cystatin_C                      : num [1:333] 9.04 9.07 8.95 9.58 8.98 ...
#>  $ EGF_R                           : num [1:333] -0.135 -0.37 -0.733 -0.422 -0.621 ...
#>  $ EN_RAGE                         : num [1:333] -3.69 -3.82 -4.76 -2.94 -2.36 ...
#>  $ ENA_78                          : num [1:333] -1.35 -1.36 -1.39 -1.37 -1.34 ...
#>  $ Eotaxin_3                       : int [1:333] 53 62 62 44 64 57 64 64 64 70 ...
#>  $ FAS                             : num [1:333] -0.0834 -0.5276 -0.6349 -0.478 -0.1278 ...
#>  $ FSH_Follicle_Stimulation_Hormon : num [1:333] -0.652 -1.627 -1.563 -0.59 -0.976 ...
#>  $ Fas_Ligand                      : num [1:333] 3.1 2.98 1.36 2.54 4.04 ...
#>  $ Fatty_Acid_Binding_Protein      : num [1:333] 2.521 2.248 0.906 0.624 2.635 ...
#>  $ Ferritin                        : num [1:333] 3.33 3.93 3.18 3.14 2.69 ...
#>  $ Fetuin_A                        : num [1:333] 1.281 1.194 1.411 0.742 2.152 ...
#>  $ Fibrinogen                      : num [1:333] -7.04 -8.05 -7.2 -7.8 -6.98 ...
#>  $ GRO_alpha                       : num [1:333] 1.38 1.37 1.41 1.37 1.4 ...
#>  $ Gamma_Interferon_induced_Monokin: num [1:333] 2.95 2.72 2.76 2.89 2.85 ...
#>  $ Glutathione_S_Transferase_alpha : num [1:333] 1.064 0.867 0.889 0.708 1.236 ...
#>  $ HB_EGF                          : num [1:333] 6.56 8.75 7.75 5.95 7.25 ...
#>  $ HCC_4                           : num [1:333] -3.04 -4.07 -3.65 -3.82 -3.15 ...
#>  $ Hepatocyte_Growth_Factor_HGF    : num [1:333] 0.5878 0.5306 0.0953 0.4055 0.5306 ...
#>  $ I_309                           : num [1:333] 3.43 3.14 2.4 3.37 3.76 ...
#>  $ ICAM_1                          : num [1:333] -0.1908 -0.462 -0.462 -0.8573 0.0972 ...
#>  $ IGF_BP_2                        : num [1:333] 5.61 5.35 5.18 5.42 5.42 ...
#>  $ IL_11                           : num [1:333] 5.12 4.94 4.67 6.22 7.07 ...
#>  $ IL_13                           : num [1:333] 1.28 1.27 1.27 1.31 1.31 ...
#>  $ IL_16                           : num [1:333] 4.19 2.88 2.62 2.44 4.74 ...
#>  $ IL_17E                          : num [1:333] 5.73 6.71 4.15 4.7 4.2 ...
#>  $ IL_1alpha                       : num [1:333] -6.57 -8.05 -8.18 -7.6 -6.94 ...
#>  $ IL_3                            : num [1:333] -3.24 -3.91 -4.65 -4.27 -3 ...
#>  $ IL_4                            : num [1:333] 2.48 2.4 1.82 1.48 2.71 ...
#>  $ IL_5                            : num [1:333] 1.099 0.693 -0.248 0.788 1.163 ...
#>  $ IL_6                            : num [1:333] 0.2694 0.0962 0.1857 -0.3712 -0.072 ...
#>  $ IL_6_Receptor                   : num [1:333] 0.6428 0.4312 0.0967 0.5752 0.0967 ...
#>  $ IL_7                            : num [1:333] 4.81 3.71 1.01 2.34 4.29 ...
#>  $ IL_8                            : num [1:333] 1.71 1.68 1.69 1.72 1.76 ...
#>  $ IP_10_Inducible_Protein_10      : num [1:333] 6.24 5.69 5.05 5.6 6.37 ...
#>  $ IgA                             : num [1:333] -6.81 -6.38 -6.32 -7.62 -4.65 ...
#>  $ Insulin                         : num [1:333] -0.626 -0.943 -1.447 -1.485 -0.3 ...
#>  $ Kidney_Injury_Molecule_1_KIM_1  : num [1:333] -1.2 -1.2 -1.19 -1.23 -1.16 ...
#>  $ LOX_1                           : num [1:333] 1.7 1.53 1.16 1.22 1.36 ...
#>  $ Leptin                          : num [1:333] -1.529 -1.466 -1.662 -1.269 -0.915 ...
#>  $ Lipoprotein_a                   : num [1:333] -4.27 -4.93 -5.84 -4.99 -2.94 ...
#>  $ MCP_1                           : num [1:333] 6.74 6.85 6.77 6.78 6.72 ...
#>  $ MCP_2                           : num [1:333] 1.981 1.809 0.401 1.981 2.221 ...
#>  $ MIF                             : num [1:333] -1.24 -1.9 -2.3 -1.66 -1.9 ...
#>  $ MIP_1alpha                      : num [1:333] 4.97 3.69 4.05 4.93 6.45 ...
#>  $ MIP_1beta                       : num [1:333] 3.26 3.14 2.4 3.22 3.53 ...
#>  $ MMP_2                           : num [1:333] 4.48 3.78 2.87 2.97 3.69 ...
#>  $ MMP_3                           : num [1:333] -2.21 -2.47 -2.3 -1.77 -1.56 ...
#>  $ MMP10                           : num [1:333] -3.27 -3.65 -2.73 -4.07 -2.62 ...
#>  $ MMP7                            : num [1:333] -3.774 -5.968 -4.03 -6.856 -0.222 ...
#>  $ Myoglobin                       : num [1:333] -1.897 -0.755 -1.386 -1.139 -1.772 ...
#>  $ NT_proBNP                       : num [1:333] 4.55 4.22 4.25 4.11 4.47 ...
#>  $ NrCAM                           : num [1:333] 5 5.21 4.74 4.97 5.2 ...
#>  $ Osteopontin                     : num [1:333] 5.36 6 5.02 5.77 5.69 ...
#>  $ PAI_1                           : num [1:333] 1.0035 -0.0306 0.4384 0 0.2523 ...
#>  $ PAPP_A                          : num [1:333] -2.9 -2.81 -2.94 -2.79 -2.94 ...
#>  $ PLGF                            : num [1:333] 4.44 4.03 4.51 3.43 4.8 ...
#>  $ PYY                             : num [1:333] 3.22 3.14 2.89 2.83 3.66 ...
#>  $ Pancreatic_polypeptide          : num [1:333] 0.579 0.336 -0.892 -0.821 0.262 ...
#>  $ Prolactin                       : num [1:333] 0 -0.5108 -0.1393 -0.0408 0.1823 ...
#>  $ Prostatic_Acid_Phosphatase      : num [1:333] -1.62 -1.74 -1.64 -1.74 -1.7 ...
#>  $ Protein_S                       : num [1:333] -1.78 -2.46 -2.26 -2.7 -1.66 ...
#>  $ Pulmonary_and_Activation_Regulat: num [1:333] -0.844 -2.303 -1.661 -1.109 -0.562 ...
#>  $ RANTES                          : num [1:333] -6.21 -6.94 -6.65 -5.99 -6.32 ...
#>  $ Resistin                        : num [1:333] -16.5 -16 -16.5 -13.5 -11.1 ...
#>   [list output truncated]