Compressive strength of concrete mixtures

Yeh (2006) describes an aggregated data set for experimental designs used to test the compressive strength of concrete mixtures. The data are used by Kuhn and Johnson (2013).

Source

Yeh I (2006). "Analysis of Strength of Concrete Using Design of Experiments and Neural Networks." Journal of Materials in Civil Engineering, 18, 597-604.

Kuhn, M., Johnson, K. (2013) Applied Predictive Modeling, Springer.

Value

concrete

a tibble

Examples

data(concrete)