Credit data

These data are from the website of Dr. Lluís A. Belanche Muñoz by way of a github repository of Dr. Gaston Sanchez. One data point is a missing outcome was removed from the original data.

Source

https://github.com/gastonstat/CreditScoring, http://bit.ly/2kkBFrk

Value

credit_data

a data frame

Examples

data(credit_data)
str(credit_data)
#> 'data.frame':	4454 obs. of  14 variables:
#>  $ Status   : Factor w/ 2 levels "bad","good": 2 2 1 2 2 2 2 2 2 1 ...
#>  $ Seniority: int  9 17 10 0 0 1 29 9 0 0 ...
#>  $ Home     : Factor w/ 6 levels "ignore","other",..: 6 6 3 6 6 3 3 4 3 4 ...
#>  $ Time     : int  60 60 36 60 36 60 60 12 60 48 ...
#>  $ Age      : int  30 58 46 24 26 36 44 27 32 41 ...
#>  $ Marital  : Factor w/ 5 levels "divorced","married",..: 2 5 2 4 4 2 2 4 2 2 ...
#>  $ Records  : Factor w/ 2 levels "no","yes": 1 1 2 1 1 1 1 1 1 1 ...
#>  $ Job      : Factor w/ 4 levels "fixed","freelance",..: 2 1 2 1 1 1 1 1 2 4 ...
#>  $ Expenses : int  73 48 90 63 46 75 75 35 90 90 ...
#>  $ Income   : int  129 131 200 182 107 214 125 80 107 80 ...
#>  $ Assets   : int  0 0 3000 2500 0 3500 10000 0 15000 0 ...
#>  $ Debt     : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ Amount   : int  800 1000 2000 900 310 650 1600 200 1200 1200 ...
#>  $ Price    : int  846 1658 2985 1325 910 1645 1800 1093 1957 1468 ...