Predicting hepatic injury from chemical information

A quantitative structure-activity relationship (QSAR) data set to predict when a molecule has risk associated with liver function.

Source

Kuhn, Max, and Kjell Johnson. Applied predictive modeling. New York: Springer, 2013.

Value

hepatic_injury_qsar

a tibble

Details

This data set was used to develop a model for predicting compounds' probability of causing hepatic injury (i.e. liver damage). This data set consisted of 281 unique compounds; 376 predictors were measured or computed for each. The response was categorical (either "none", "mild", or "severe"), and was highly unbalanced.

This kind of response often occurs in pharmaceutical data because companies steer away from creating molecules that have undesirable characteristics. Therefore, well-behaved molecules often greatly outnumber undesirable molecules. The predictors consisted of measurements from 184 biological screens and 192 chemical feature predictors. The biological predictors represent activity for each screen and take values between 0 and 10 with a mode of 4. The chemical feature predictors represent counts of important sub-structures as well as measures of physical properties that are thought to be associated with hepatic injury.

Columns:

class: ordered and factor (levels: 'none', 'mild', and 'severe')

bio_assay_001 - bio_assay_184: numeric

chem_fp_001 - chem_fp_192: numeric

Examples

data(hepatic_injury_qsar)
str(hepatic_injury_qsar)
#> tibble [281 × 377] (S3: tbl_df/tbl/data.frame)
#>  $ class        : Ord.factor w/ 3 levels "none"<"mild"<..: 2 1 1 3 2 3 1 1 1 1 ...
#>  $ bio_assay_001: num [1:281] 4 4 4 4.9 4 4 4 4 4 4 ...
#>  $ bio_assay_002: num [1:281] 4 4.5 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_003: num [1:281] 4 4 4 4.5 4 4 4 4.5 4 4 ...
#>  $ bio_assay_004: num [1:281] 6 6 6 4 6 6 6 6 5.2 6 ...
#>  $ bio_assay_005: num [1:281] 4 4 4 4 4 4 4 4 4.5 4 ...
#>  $ bio_assay_006: num [1:281] 4 4 4 4.9 4 4 4.5 4 4.5 4 ...
#>  $ bio_assay_007: num [1:281] 4 4 4 4.5 4 4 4.5 4 4 4.5 ...
#>  $ bio_assay_008: num [1:281] 4 4 4 4 4 4 4 4 5 4 ...
#>  $ bio_assay_009: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_010: num [1:281] 4 4.5 4.5 4 4 4 4 4.5 5.3 4 ...
#>  $ bio_assay_011: num [1:281] 4 4 4 4 4 4 4.5 4 4.5 4 ...
#>  $ bio_assay_012: num [1:281] 4 4 4 4 4.5 4.5 4 4.5 4.5 4 ...
#>  $ bio_assay_013: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_014: num [1:281] 4 4 4 4 4 4 4 4 5.2 4 ...
#>  $ bio_assay_015: num [1:281] 4 4 4 6.9 5.5 4.7 4.5 4 4.5 5.3 ...
#>  $ bio_assay_016: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_017: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_018: num [1:281] 4 4 4 4.9 4 4 4.5 4 4.5 4 ...
#>  $ bio_assay_019: num [1:281] 4 4.5 4 5.5 4 4 6.5 4 6.5 4 ...
#>  $ bio_assay_020: num [1:281] 4 4.5 4 4 4.5 4 4 4.5 4 4 ...
#>  $ bio_assay_021: num [1:281] 4 4 4 4 4.5 4 4.5 4 4 4 ...
#>  $ bio_assay_022: num [1:281] 4 4 4 4 4 4 4 0 4 4 ...
#>  $ bio_assay_023: num [1:281] 4 4 4 4 4 4 4 4 4.5 4 ...
#>  $ bio_assay_024: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_025: num [1:281] 4 4 4 4.8 4 4 4 4 5.3 4 ...
#>  $ bio_assay_026: num [1:281] 4 4 4 5.2 4 4 4 4 5.4 4 ...
#>  $ bio_assay_027: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_028: num [1:281] 4.5 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_029: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_030: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_031: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_032: num [1:281] 4 4 4 0 4 4 4 4 0 4 ...
#>  $ bio_assay_033: int [1:281] 4 0 4 4 4 4 4 4 0 4 ...
#>  $ bio_assay_034: num [1:281] 4 4 4 6 4 4 5.9 4 6.3 4 ...
#>  $ bio_assay_035: num [1:281] 4 4 4 6.3 4 4 4 4 5.7 4 ...
#>  $ bio_assay_036: num [1:281] 4 4 4 4 5.8 4 5.2 4 7.5 4 ...
#>  $ bio_assay_037: num [1:281] 4 4 4 4.5 4 4 4.8 4 6.1 4 ...
#>  $ bio_assay_038: num [1:281] 4 4 4 5.8 4 4.5 4 4 4.5 4 ...
#>  $ bio_assay_039: num [1:281] 4 4 4.5 4.5 4 4 4 4 4 4 ...
#>  $ bio_assay_040: num [1:281] 4.5 4 4 5.5 4 4 6.1 5.3 5.5 4.5 ...
#>  $ bio_assay_041: num [1:281] 4 4 4 4 4 4 4 4 5 4 ...
#>  $ bio_assay_042: num [1:281] 4 4 4.5 4 4 4 5.6 4 4.5 4 ...
#>  $ bio_assay_043: num [1:281] 4 4 4 4.5 4 4 4 4 4 4 ...
#>  $ bio_assay_044: num [1:281] 4 4 4 6 4 4 4.5 4 4 4 ...
#>  $ bio_assay_045: num [1:281] 4 4 4 4 4 0 4 4 4 4 ...
#>  $ bio_assay_046: num [1:281] 4 4 4 5.1 4 4 4.5 4 5.7 4 ...
#>  $ bio_assay_047: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_048: num [1:281] 4 4 4 4.9 4 4 4.5 4 5.7 4 ...
#>  $ bio_assay_049: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_050: num [1:281] 4 4 4 4.5 4 4 4 4 4.9 4 ...
#>  $ bio_assay_051: num [1:281] 4 4 4 5 4 4 5.4 4 7.2 4 ...
#>  $ bio_assay_052: num [1:281] 4 4 4 4.7 4 4 4.9 4 6.2 4 ...
#>  $ bio_assay_053: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_054: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_055: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_056: num [1:281] 4 4 4 4 4 0 4 4 4 4 ...
#>  $ bio_assay_057: num [1:281] 4 4 4 4 4 0 4 4 4 4 ...
#>  $ bio_assay_058: num [1:281] 4 4 4 4 4 4 4 4 5.4 4 ...
#>  $ bio_assay_059: num [1:281] 4 4 4 4 4 4 4 4 4 4.5 ...
#>  $ bio_assay_060: num [1:281] 4 4 4 4 4 4 4 4 5.1 4 ...
#>  $ bio_assay_061: num [1:281] 4 4 4 4 4 4 4 4 0 4 ...
#>  $ bio_assay_062: num [1:281] 4 4 4 4 4 4 4 4 4.5 4 ...
#>  $ bio_assay_063: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_064: num [1:281] 4 4 4 5.3 4 4 5.1 4 6.8 4 ...
#>  $ bio_assay_065: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_066: num [1:281] 4 4 4 4 4 4 4 4 0 4 ...
#>  $ bio_assay_067: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_068: num [1:281] 4 4 4 4.6 4 4 5.6 4 5.9 4.5 ...
#>  $ bio_assay_069: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_070: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_071: num [1:281] 4 4 4 5.5 4 4 5 4 6 4 ...
#>  $ bio_assay_072: num [1:281] 4 4 4 5.5 4 4 4 4 5.7 4 ...
#>  $ bio_assay_073: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_074: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_075: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_076: num [1:281] 4 4 4 4 4 4 4 4.5 4 4 ...
#>  $ bio_assay_077: num [1:281] 4 4 4 4.5 4 4 4 4 4.5 4 ...
#>  $ bio_assay_078: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_079: num [1:281] 4 4 4 4 4 4 4 4 4.2 4 ...
#>  $ bio_assay_080: num [1:281] 4 4 4 0 4 4.5 4 4 4 4 ...
#>  $ bio_assay_081: num [1:281] 4 0 4 0 4 4 4 4 4 4.5 ...
#>  $ bio_assay_082: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_083: num [1:281] 4.5 4 4 4 4 4 4 4.5 4.5 4 ...
#>  $ bio_assay_084: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_085: num [1:281] 5.4 4 4 4.5 4.5 4 4 4 4 4 ...
#>  $ bio_assay_086: num [1:281] 4 4 4 4.9 4 4 4 4 5.2 4 ...
#>  $ bio_assay_087: num [1:281] 4 4 4 4.5 4 4 4.5 4 4 4 ...
#>  $ bio_assay_088: int [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_089: num [1:281] 4 4 4 4.5 4 4 4 4 4 4 ...
#>  $ bio_assay_090: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_091: num [1:281] 4 4 4 5.3 4 4 4 4 4 4 ...
#>  $ bio_assay_092: num [1:281] 4 4 4 4.1 4 4 4 4 4 4 ...
#>  $ bio_assay_093: num [1:281] 4 4 4 4 4 0 4 4 4 4 ...
#>  $ bio_assay_094: num [1:281] 4 4 4 4 4 0 4 4 4 4 ...
#>  $ bio_assay_095: num [1:281] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ bio_assay_096: num [1:281] 4.5 4 4 4 4 4 4 4 5 4 ...
#>  $ bio_assay_097: num [1:281] 4 4.4 4 5.1 4.7 4 7.2 4 5.4 5.1 ...
#>  $ bio_assay_098: num [1:281] 4 4 4 6.1 4.5 4 6.2 4 4 4.5 ...
#>   [list output truncated]