Parkinson's disease speech classification data set

Parkinson's disease speech classification data set

Source

UCI ML repository (data) https://archive.ics.uci.edu/ml/datasets/Parkinson%27s+Disease+Classification#,

Sakar et al (2019), "A comparative analysis of speech signal processing algorithms for Parkinson’s disease classification and the use of the tunable Q-factor wavelet transform", Applied Soft Computing, V74, pg 255-263.

Value

pd_speech

a data frame

Details

From the UCI ML archive, the description is "The data used in this study were gathered from 188 patients with PD (107 men and 81 women) with ages ranging from 33 to 87 (65.1 p/m 10.9) at the Department of Neurology in Cerrahpaşa Faculty of Medicine, Istanbul University. The control group consists of 64 healthy individuals (23 men and 41 women) with ages varying between 41 and 82 (61.1 p/m 8.9). During the data collection process, the microphone is set to 44.1 KHz and following the physician's examination, the sustained phonation of the vowel /a/ was collected from each subject with three repetitions."

The data here are averaged over the replicates.

Examples

data(pd_speech)
str(pd_speech)
#> tibble [252 × 752] (S3: tbl_df/tbl/data.frame)
#>  $ gender                        : num [1:252] 1 0 1 0 0 1 1 1 1 1 ...
#>  $ PPE                           : num [1:252] 0.823 0.416 0.802 0.829 0.831 ...
#>  $ DFA                           : num [1:252] 0.696 0.794 0.62 0.626 0.779 ...
#>  $ RPDE                          : num [1:252] 0.567 0.592 0.521 0.537 0.727 ...
#>  $ meanPeriodPulses              : num [1:252] 0.00822 0.00888 0.00604 0.00391 0.00562 ...
#>  $ stdDevPeriodPulses            : num [1:252] 7.34e-05 1.85e-03 1.04e-04 4.21e-05 2.02e-03 ...
#>  $ locPctJitter                  : num [1:252] 0.001963 0.00579 0.002217 0.000757 0.003593 ...
#>  $ locAbsJitter                  : num [1:252] 1.61e-05 4.96e-05 1.34e-05 2.97e-06 2.10e-05 ...
#>  $ rapJitter                     : num [1:252] 5.87e-04 1.82e-03 3.23e-04 8.67e-05 1.10e-03 ...
#>  $ ppq5Jitter                    : num [1:252] 0.001173 0.00269 0.000833 0.00024 0.002057 ...
#>  $ ddpJitter                     : num [1:252] 0.00176 0.005473 0.000973 0.00026 0.00329 ...
#>  $ locShimmer                    : num [1:252] 0.071 0.0627 0.041 0.0495 0.1592 ...
#>  $ locDbShimmer                  : num [1:252] 0.639 0.567 0.354 0.457 1.454 ...
#>  $ apq3Shimmer                   : num [1:252] 0.0348 0.0304 0.0196 0.0257 0.0773 ...
#>  $ apq5Shimmer                   : num [1:252] 0.0456 0.0365 0.0261 0.0275 0.1084 ...
#>  $ apq11Shimmer                  : num [1:252] 0.0626 0.054 0.0414 0.0443 0.1654 ...
#>  $ ddaShimmer                    : num [1:252] 0.1043 0.0911 0.0588 0.0772 0.232 ...
#>  $ meanAutoCorrHarmonicity       : num [1:252] 0.977 0.95 0.988 0.992 0.896 ...
#>  $ meanNoiseToHarmHarmonicity    : num [1:252] 0.02684 0.06874 0.01202 0.00851 0.12841 ...
#>  $ meanHarmToNoiseHarmonicity    : num [1:252] 19.4 18 20.4 23.5 10.8 ...
#>  $ minIntensity                  : num [1:252] 66.7 75.7 75.1 70.9 57.4 ...
#>  $ maxIntensity                  : num [1:252] 73.6 80.9 78.4 78.9 65.4 ...
#>  $ meanIntensity                 : num [1:252] 70.7 79 77 75.6 62.4 ...
#>  $ f1                            : num [1:252] 551 833 683 712 504 ...
#>  $ f2                            : num [1:252] 1027 1203 1153 1273 1221 ...
#>  $ f3                            : num [1:252] 2396 3178 2979 2057 2071 ...
#>  $ f4                            : num [1:252] 3611 4000 3868 3203 3235 ...
#>  $ b1                            : num [1:252] 107 293 123 106 652 ...
#>  $ b2                            : num [1:252] 74.8 208.3 85.5 79.3 683.6 ...
#>  $ b3                            : num [1:252] 218 323 169 2519 715 ...
#>  $ b4                            : num [1:252] 233 525 516 325 561 ...
#>  $ GQ_prc5_95                    : num [1:252] 0.803 0.898 1 1 0.914 ...
#>  $ GQ_std_cycle_open             : num [1:252] 17.49 115.31 9.05 9.81 87.72 ...
#>  $ GQ_std_cycle_closed           : num [1:252] 2.81 18.22 0 0 10.76 ...
#>  $ GNE_mean                      : num [1:252] 1.015 0.841 0.834 1.631 0.841 ...
#>  $ GNE_std                       : num [1:252] 0.2142 0.1876 0.0799 0.3902 0.1455 ...
#>  $ GNE_SNR_TKEO                  : num [1:252] 0.1091 0.1477 0.1553 0.0672 0.1702 ...
#>  $ GNE_SNR_SEO                   : num [1:252] 1257257 1851946 1438735 1440965 1532406 ...
#>  $ GNE_NSR_TKEO                  : num [1:252] 1.59 1.67 1.65 1.47 1.69 ...
#>  $ GNE_NSR_SEO                   : num [1:252] 3.04 3.03 3.02 3.08 3.06 ...
#>  $ VFER_mean                     : num [1:252] 0.000479 0.000253 0.001704 0.003726 0.001358 ...
#>  $ VFER_std                      : num [1:252] 0.000274 0.000218 0.000507 0.001731 0.003217 ...
#>  $ VFER_entropy                  : num [1:252] 0.839 0.61 3.451 10.038 3.071 ...
#>  $ VFER_SNR_TKEO                 : num [1:252] 230.4 60.3 51.1 99.4 101.7 ...
#>  $ VFER_SNR_SEO                  : num [1:252] 407.4 141.9 392 397.5 98.5 ...
#>  $ VFER_NSR_TKEO                 : num [1:252] 1.28 1.23 1.23 1.32 1.25 ...
#>  $ VFER_NSR_SEO                  : num [1:252] 1.32 1.29 1.35 1.37 1.21 ...
#>  $ IMF_SNR_SEO                   : num [1:252] 45.251 82.136 13.799 0.417 39.002 ...
#>  $ IMF_SNR_TKEO                  : num [1:252] 7.9692 5.1628 1.5484 0.0901 1.3481 ...
#>  $ IMF_SNR_entropy               : num [1:252] 28.959 31.376 6.983 0.853 12.213 ...
#>  $ IMF_NSR_SEO                   : num [1:252] 0.256 0.272 0.125 0.137 0.211 ...
#>  $ IMF_NSR_TKEO                  : num [1:252] 5.63 5.58 10.92 35.8 8.75 ...
#>  $ IMF_NSR_entropy               : num [1:252] 0.192 0.198 0.135 0.137 0.182 ...
#>  $ mean_Log_energy               : num [1:252] 9.28 9.93 9.89 9.67 9.37 ...
#>  $ mean_MFCC_0th_coef            : num [1:252] 11.7 14.1 16.6 12.3 14.7 ...
#>  $ mean_MFCC_1st_coef            : num [1:252] 8.21 6.64 6.09 6.74 4.43 ...
#>  $ mean_MFCC_2nd_coef            : num [1:252] 2.869 2.909 -0.432 0.498 3.326 ...
#>  $ mean_MFCC_3rd_coef            : num [1:252] -0.309 0.836 1.041 -2.216 -1.87 ...
#>  $ mean_MFCC_4th_coef            : num [1:252] 0.435 -1.689 -1.882 -1.762 -1.952 ...
#>  $ mean_MFCC_5th_coef            : num [1:252] -1.57 -0.969 -2.724 -1.215 0.392 ...
#>  $ mean_MFCC_6th_coef            : num [1:252] -1.56 -0.6 -1.14 -2.25 -1.29 ...
#>  $ mean_MFCC_7th_coef            : num [1:252] 0.4265 0.8811 0.8524 0.3172 -0.0881 ...
#>  $ mean_MFCC_8th_coef            : num [1:252] 0.6196 1.5336 0.0925 1.323 0.6159 ...
#>  $ mean_MFCC_9th_coef            : num [1:252] 0.1647 -0.0641 0.9851 -1.2218 -1.1646 ...
#>  $ mean_MFCC_10th_coef           : num [1:252] 0.63 -1.047 -0.613 -0.937 -0.687 ...
#>  $ mean_MFCC_11th_coef           : num [1:252] 0.113 -0.872 -1.08 -0.478 -1.556 ...
#>  $ mean_MFCC_12th_coef           : num [1:252] -0.5713 0.0172 0.4279 -0.0895 -0.7913 ...
#>  $ mean_delta_log_energy         : num [1:252] -5.91e-04 1.53e-03 8.03e-05 -3.50e-03 -6.45e-03 ...
#>  $ mean_0th_delta                : num [1:252] -0.003648 -0.006824 0.000593 -0.010634 -0.00694 ...
#>  $ mean_1st_delta                : num [1:252] -0.00115 -0.00172 -0.00221 0.00256 -0.00728 ...
#>  $ mean_2nd_delta                : num [1:252] 0.00338 0.00645 0.00438 0.00388 0.00233 ...
#>  $ mean_3rd_delta                : num [1:252] -2.89e-04 5.94e-04 1.10e-04 -2.17e-06 1.25e-03 ...
#>  $ mean_4th_delta                : num [1:252] -0.0005 -0.001445 -0.000689 0.001075 -0.003117 ...
#>  $ mean_5th_delta                : num [1:252] -0.000543 0.001018 -0.001211 -0.000138 0.00282 ...
#>  $ mean_6th_delta                : num [1:252] 6.86e-04 1.95e-03 6.32e-04 -1.31e-03 9.87e-06 ...
#>  $ mean_7th_delta                : num [1:252] 0.000618 -0.002802 -0.000874 0.000407 -0.000266 ...
#>  $ mean_8th_delta                : num [1:252] 0.000673 -0.002118 0.000747 -0.000263 0.002366 ...
#>  $ mean_9th_delta                : num [1:252] -0.000494 0.002002 0.00111 0.00213 -0.000315 ...
#>  $ mean_10th_delta               : num [1:252] -0.000101 -0.001611 -0.000585 0.002704 0.001459 ...
#>  $ mean_11th_delta               : num [1:252] 0.00132 -0.001233 0.000134 -0.000799 0.00243 ...
#>  $ mean_12th_delta               : num [1:252] 0.000282 -0.001075 0.00133 -0.00005 -0.001658 ...
#>  $ mean_delta_delta_log_energy   : num [1:252] 2.19e-04 -8.21e-05 -5.44e-06 1.66e-04 -1.43e-04 ...
#>  $ mean_delta_delta_0th          : num [1:252] -2.61e-04 -2.09e-04 -1.64e-05 -1.29e-04 -3.83e-04 ...
#>  $ mean_1st_delta_delta          : num [1:252] 4.07e-04 -2.73e-04 1.75e-04 -1.55e-04 -3.27e-05 ...
#>  $ mean_2nd_delta_delta          : num [1:252] 0.00047 0.000106 -0.000545 -0.000154 0.000322 ...
#>  $ mean_3rd_delta_delta          : num [1:252] -4.51e-04 1.29e-04 4.82e-05 9.82e-05 2.42e-04 ...
#>  $ mean_4th_delta_delta          : num [1:252] -2.43e-04 1.65e-04 1.93e-04 -2.55e-04 6.73e-05 ...
#>  $ mean_5th_delta_delta          : num [1:252] 7.08e-05 -5.77e-05 -1.57e-04 -8.74e-05 1.97e-04 ...
#>  $ mean_6th_delta_delta          : num [1:252] -4.08e-04 -1.01e-04 -2.64e-04 -3.72e-05 2.54e-04 ...
#>  $ mean_7th_delta_delta          : num [1:252] -2.83e-04 1.58e-04 2.79e-04 7.31e-05 1.95e-05 ...
#>  $ mean_8th_delta_delta          : num [1:252] 1.07e-04 9.06e-05 -1.48e-04 9.80e-05 2.36e-04 ...
#>  $ mean_9th_delta_delta          : num [1:252] 1.88e-04 1.59e-04 5.00e-05 2.45e-05 2.24e-04 ...
#>  $ mean_10th_delta_delta         : num [1:252] -4.26e-05 2.73e-04 -1.03e-04 -2.02e-04 3.81e-04 ...
#>  $ mean_11th_delta_delta         : num [1:252] 1.49e-04 6.84e-05 -8.47e-05 -2.46e-04 4.69e-05 ...
#>  $ mean_12th_delta_delta         : num [1:252] -1.27e-04 -8.65e-05 2.30e-05 -2.36e-04 -5.54e-05 ...
#>  $ std_Log_energy                : num [1:252] 0.366 0.243 0.156 0.407 0.432 ...
#>  $ std_MFCC_0th_coef             : num [1:252] 0.956 1.018 0.618 1.137 0.698 ...
#>  $ std_MFCC_1st_coef             : num [1:252] 0.575 0.584 0.403 0.485 0.48 ...
#>  $ std_MFCC_2nd_coef             : num [1:252] 0.723 0.558 0.389 0.463 0.476 ...
#>   [list output truncated]