Morphometric data on scat

Reid (2015) collected data on animal feses in coastal California. The data consist of DNA verified species designations as well as fields related to the time and place of the collection and the scat itself. The data are on the three main species.

Source

Reid, R. E. B. (2015). A morphometric modeling approach to distinguishing among bobcat, coyote and gray fox scats. Wildlife Biology, 21(5), 254-262

Value

scat

a tibble

Examples

data(scat)
str(scat)
#> tibble [110 × 19] (S3: tbl_df/tbl/data.frame)
#>  $ Species  : Factor w/ 3 levels "bobcat","coyote",..: 2 2 1 2 2 2 1 1 1 1 ...
#>  $ Month    : Factor w/ 9 levels "April","August",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ Year     : int [1:110] 2012 2012 2012 2012 2012 2012 2012 2012 2012 2012 ...
#>  $ Site     : Factor w/ 2 levels "ANNU","YOLA": 2 2 2 2 2 2 1 1 1 1 ...
#>  $ Location : Factor w/ 3 levels "edge","middle",..: 1 1 2 2 1 1 3 3 3 2 ...
#>  $ Age      : int [1:110] 5 3 3 5 5 5 1 3 5 5 ...
#>  $ Number   : int [1:110] 2 2 2 2 4 3 5 7 2 1 ...
#>  $ Length   : num [1:110] 9.5 14 9 8.5 8 9 6 5.5 11 20.5 ...
#>  $ Diameter : num [1:110] 25.7 25.4 18.8 18.1 20.7 21.2 15.7 21.9 17.5 18 ...
#>  $ Taper    : num [1:110] 41.9 37.1 16.5 24.7 20.1 28.5 8.2 19.3 29.1 21.4 ...
#>  $ TI       : num [1:110] 1.63 1.46 0.88 1.36 0.97 1.34 0.52 0.88 1.66 1.19 ...
#>  $ Mass     : num [1:110] 15.9 17.6 8.4 7.4 25.4 ...
#>  $ d13C     : num [1:110] -26.9 -29.6 -28.7 -20.1 -23.2 ...
#>  $ d15N     : num [1:110] 6.94 9.87 8.52 5.79 7.01 8.28 4.2 3.89 7.34 6.06 ...
#>  $ CN       : num [1:110] 8.5 11.3 8.1 11.5 10.6 9 5.4 5.6 5.8 7.7 ...
#>  $ ropey    : int [1:110] 0 0 1 1 0 1 1 0 0 1 ...
#>  $ segmented: int [1:110] 0 0 1 0 1 0 1 1 1 1 ...
#>  $ flat     : int [1:110] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ scrape   : int [1:110] 0 0 1 0 0 0 1 0 0 0 ...