Kelly Blue Book resale data for 2005 model year GM cars

Kuiper (2008) collected data on Kelly Blue Book resale data for 804 GM cars (2005 model year).

Source

Kuiper, S. (2008). Introduction to Multiple Regression: How Much Is Your Car Worth?, Journal of Statistics Education, Vol. 16 http://jse.amstat.org/jse_archive.htm#2008.

Value

car_prices

data frame of the suggested retail price (column Price) and various characteristics of each car (columns Mileage, Cylinder, Doors, Cruise, Sound, Leather, Buick, Cadillac, Chevy, Pontiac, Saab, Saturn, convertible, coupe, hatchback, sedan and wagon)

Examples

data(car_prices)
str(car_prices)
#> tibble [804 × 18] (S3: tbl_df/tbl/data.frame)
#>  $ Price      : num [1:804] 22661 21725 29143 30732 33359 ...
#>  $ Mileage    : int [1:804] 20105 13457 31655 22479 17590 23635 17381 27558 25049 17319 ...
#>  $ Cylinder   : int [1:804] 6 6 4 4 4 4 4 4 4 4 ...
#>  $ Doors      : int [1:804] 4 2 2 2 2 2 2 2 2 4 ...
#>  $ Cruise     : int [1:804] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ Sound      : int [1:804] 0 1 1 0 1 0 1 0 0 0 ...
#>  $ Leather    : int [1:804] 0 0 1 0 1 0 1 1 0 1 ...
#>  $ Buick      : int [1:804] 1 0 0 0 0 0 0 0 0 0 ...
#>  $ Cadillac   : int [1:804] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ Chevy      : int [1:804] 0 1 0 0 0 0 0 0 0 0 ...
#>  $ Pontiac    : int [1:804] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ Saab       : int [1:804] 0 0 1 1 1 1 1 1 1 1 ...
#>  $ Saturn     : int [1:804] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ convertible: int [1:804] 0 0 1 1 1 1 1 1 1 0 ...
#>  $ coupe      : int [1:804] 0 1 0 0 0 0 0 0 0 0 ...
#>  $ hatchback  : int [1:804] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ sedan      : int [1:804] 1 0 0 0 0 0 0 0 0 1 ...
#>  $ wagon      : int [1:804] 0 0 0 0 0 0 0 0 0 0 ...