Watson churn data

Watson churn data

Source

IBM Watson Analytics https://ibm.co/2sOvyvy

Value

wa_churn

a data frame

Details

These data were downloaded from the IBM Watson site (see below) in September 2018. The data contain a factor for whether a customer churned or not. Alternatively, the tenure column presumably contains information on how long the customer has had an account. A survival analysis can be done on this column using the churn outcome as the censoring information. A data dictionary can be found on the source website.

Examples

data(wa_churn)
str(wa_churn)
#> tibble [7,043 × 20] (S3: tbl_df/tbl/data.frame)
#>  $ churn            : Factor w/ 2 levels "Yes","No": 2 2 1 2 1 1 2 2 1 2 ...
#>  $ female           : num [1:7043] 1 0 0 0 1 1 0 1 1 0 ...
#>  $ senior_citizen   : int [1:7043] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ partner          : num [1:7043] 1 0 0 0 0 0 0 0 1 0 ...
#>  $ dependents       : num [1:7043] 0 0 0 0 0 0 1 0 0 1 ...
#>  $ tenure           : int [1:7043] 1 34 2 45 2 8 22 10 28 62 ...
#>  $ phone_service    : num [1:7043] 0 1 1 0 1 1 1 0 1 1 ...
#>  $ multiple_lines   : Factor w/ 3 levels "No","No phone service",..: 2 1 1 2 1 3 3 2 3 1 ...
#>  $ internet_service : Factor w/ 3 levels "DSL","Fiber optic",..: 1 1 1 1 2 2 2 1 2 1 ...
#>  $ online_security  : Factor w/ 3 levels "No","No internet service",..: 1 3 3 3 1 1 1 3 1 3 ...
#>  $ online_backup    : Factor w/ 3 levels "No","No internet service",..: 3 1 3 1 1 1 3 1 1 3 ...
#>  $ device_protection: Factor w/ 3 levels "No","No internet service",..: 1 3 1 3 1 3 1 1 3 1 ...
#>  $ tech_support     : Factor w/ 3 levels "No","No internet service",..: 1 1 1 3 1 1 1 1 3 1 ...
#>  $ streaming_tv     : Factor w/ 3 levels "No","No internet service",..: 1 1 1 1 1 3 3 1 3 1 ...
#>  $ streaming_movies : Factor w/ 3 levels "No","No internet service",..: 1 1 1 1 1 3 1 1 3 1 ...
#>  $ contract         : Factor w/ 3 levels "Month-to-month",..: 1 2 1 2 1 1 1 1 1 2 ...
#>  $ paperless_billing: num [1:7043] 1 0 1 0 1 1 1 0 1 0 ...
#>  $ payment_method   : Factor w/ 4 levels "Bank transfer (automatic)",..: 3 4 4 1 3 3 2 4 3 1 ...
#>  $ monthly_charges  : num [1:7043] 29.9 57 53.9 42.3 70.7 ...
#>  $ total_charges    : num [1:7043] 29.9 1889.5 108.2 1840.8 151.7 ...